ecorce_mastersheets/requirements.txt
# requirements.txt (lean, pinned)
Flask==2.3.2
gunicorn==21.2.0
python-dotenv==1.0.1
stripe==7.11.0
reportlab==4.2.5

# Gemini (public API)
google-genai==0.3.0

# keep protobuf/grpc reasonable
protobuf<6,>=4.25.3
grpcio>=1.62.0,<2
google-api-core>=2.17.0,<3
google-generativeai>=0.7,<1